+++ packages/components/src/Form/Fields/FieldTimeSelect/FieldTimeSelect.story.tsx
@@ -93,6 +93,7 @@ export const Controlled = () => {
           label="Standard Time"
           value={controlledTime}
           onChange={setControlledTime}
+          autoFocus
         />
         <FieldTimeSelect
           label="Military Time"